fix: vscode refactors not good
