+++ packages/shared/comms/handlers.ts
@@ -29,7 +29,7 @@ import { handleCommsDisconnection } from './actions'
 import { Avatar } from '@dcl/schemas'
 import { Observable } from 'mz-observable'
 import { eventChannel } from 'redux-saga'
-import { ensureFriendProfile } from 'shared/profiles/ProfileAsPromise'
+import { ensureFriendProfile } from 'shared/profiles/ensureFriendsProfile'
 import { trackEvent } from 'shared/analytics'
 import { ProfileType } from 'shared/profiles/types'
 import { ensureAvatarCompatibilityFormat } from 'shared/profiles/transformations/profileToServerFormat'
