+++ packages/unity-interface/fetchENSOwnerProfile.ts
@@ -2,7 +2,7 @@ import { isAddress } from 'eth-connect'
 import { ethereumConfigurations } from 'config'
 import { fetchENSOwnersContains, getAppNetwork } from 'shared/web3'
 import { ProfileType } from 'shared/profiles/types'
-import { ensureFriendProfile } from 'shared/profiles/ProfileAsPromise'
+import { ensureFriendProfile } from 'shared/profiles/ensureFriendsProfile'
 import { Avatar } from '@dcl/schemas'
 
 export async function fetchENSOwnerProfile(name: string, maxResults: number = 1): Promise<Avatar[]> {
